requirements.txt
Django==1.6.1
South==0.8.4
distribute==0.6.49
python-gcm==0.1.4
requests==2.1.0
stevedore==0.12
virtualenv==1.9.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.1.1
wsgiref==0.1.2
